requirements.txt
django
djangorestframework
django-cors-headers
requests
Pillow
gunicorn
whitenoise
